Create Note/Notification sections added
