+++ components/Sidebar/Sidebar.module.css
@@ -3,14 +3,12 @@
   flex-direction: column;
   align-items: center;
   height: 100%;
-  margin-bottom: 5rem;
   width: 17vw;
   border: 1px solid var(--primary-5);
   border-radius: 25px;
-  margin-top: 2vh;
+  margin-top: 4vh;
   margin-left: 2vw;
   padding-bottom: 2rem;
-  padding-right: 2rem;
   background-color: var(--primary-5);
 }
 .sidebarDevslogo img {
@@ -34,9 +32,6 @@
   display: flex;
   /* align-items: center; */
   justify-content: center;
-  margin-left: 1rem;
-  margin-right: auto;
-  margin-top: 1rem;
   margin-top: 1rem;
   width: 100%;
 }
@@ -46,7 +41,6 @@
   height: 50px;
   border-radius: 50%;
   margin-top: 1.5rem;
-  margin-right: 2rem;
 }
 
 .sidebarUser {
@@ -193,10 +187,6 @@
     margin: 2vh 2vw;
   }
 
-  .sidebarFunctions {
-    display: none;
-  }
-
   .sidebarMenu {
     display: block;
     margin-top: 1rem;
@@ -219,22 +209,72 @@
   .toggleCheckbox:checked ~ .sidebarHeader .sidebarUser .sidebarUsermail {
     display: none;
   }
-
-  /* .toggleIcon:before {
-    transform: translateY(6px) rotate(45deg);
-  } */
-
-  /* .toggleIcon:after {
-    transform: translateY(-6px) rotate(-45deg);
-  } */
 }
 @media screen and (max-width: 690px) {
-  .sidebarWrapper {
+  /* .sidebarWrapper {
     display: none;
-  }
+  } */
 }
 @media screen and (max-width: 1000px) {
   .sidebarWrapper {
     width: 20rem;
+  }
+}
+
+@media (max-width: 690px) {
+  /* .hamburgerDisplay {
+  display: none;
+} */
+  .hamburger-menu {
+    display: flex;
+    flex-direction: column;
+  }
+
+  .hamburger-icon {
+    width: 20px;
+    height: 20px;
+    cursor: pointer;
+    position: absolute;
+    margin-top: 2rem;
+    right: 5%;
+  }
+
+  .line {
+    width: 100%;
+    height: 3px;
+    background-color: black;
+    margin-bottom: 5px;
+  }
+
+  .open .line:nth-child(1) {
+    transform: rotate(-45deg) translate(-6px, 5px);
+  }
+
+  .open .line:nth-child(2) {
+    opacity: 0;
+  }
+
+  .open .line:nth-child(3) {
+    transform: rotate(45deg) translate(-6px, -5px);
+  }
+
+  .anchoryy {
+    display: none;
+  }
+
+  .open {
+    display: block;
+    width: 100vw;
+  }
+  .hamburgerInfo {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    margin-left: 5rem;
+  }
+}
+@media (min-width: 690px) {
+  .hamburgerDisplay {
+    display: block;
   }
 }
