+++ components/Sidebar/link.js
@@ -1,0 +1,28 @@
+const Features = [
+    {
+        mainTitle: 'Dashboard',
+        mainLink: '/Dashboard',
+        iconLink: 'assets/images/Dashboard.png'
+    },
+    {
+        mainTitle: 'Create Notifications',
+        mainLink: '/CreateNotification',
+        iconLink: 'assets/images/Bell.png'
+    },
+    {
+        mainTitle: 'Past Notifications',
+        mainLink: '/PastNotification',
+        iconLink: 'assets/images/Bell.png'
+    },
+    {
+        mainTitle: 'Create Notes',
+        mainLink: '/CreateNote',
+        iconLink: 'assets/images/Notes.png'
+    },
+    {
+        mainTitle: 'Past Notes',
+        mainLink: '/PastNote',
+        iconLink: 'assets/images/Notes.png'
+    }
+]
+export default Features
