--- components/photoassets/PhotoAssets.js
@@ -1,11 +1,0 @@
-import Bell from "../../public/assets/images/Bell.png";
-import Dashboard from "../../public/assets/images/Dashboard.png";
-import DP from "../../public/assets/images/DP.png";
-import Notes from "../../public/assets/images/Notes.png";
-
-export const photos = {
-  Bell,
-  Dashboard,
-  DP,
-  Notes,
-};
